@@ -2,7 +2,7 @@ import * as network from 'actions/network'
 import {loadState} from 'utils/storage'
 const {addresses} = CONFIG.web3
 
-const loadedState = loadState('state.network', CONFIG.web3.bridge.network)
+const loadedState = loadState('state.network') || CONFIG.web3.bridge.network
 
 const initialState = {
   addresses,
