@@ -39,19 +39,27 @@ module.exports = {
     addresses: {
       ropsten: {
         ColuLocalNetwork: '0x41C9d91E96b933b74ae21bCBb617369CBE022530',
-        CurrencyFactory: '0xA1F05144f9d3298a702c8EEE3ca360bc87d05207',
         TokenFactory: '0x824E01Cf7013f459Aa010D73627B006a8740b183',
         ForeignBridgeFactory: '0x7D913f6F0bA5Cc0660d8F60Df894b01F272550A0'
       },
       mainnet: {
         ColuLocalNetwork: '0x4162178B78D6985480A308B2190EE5517460406D',
-        CurrencyFactory: '0xE3e3bed21fC39d0915f66509eD0AAc05dB6d6454',
-        TokenFactory: '0xac051e086FD2046FC75A53D38088B4DD6e00E25b'
+        TokenFactory: '0xac051e086FD2046FC75A53D38088B4DD6e00E25b',
+        ForeignBridgeFactory: '0xE600496e0267D6b7AFDb62f83D46062199f0B0d7'
       },
-      fuse: {
-        HomeBridgeFactory: '0xFb5CC1688Ec06c57cbAB6cC34c33413154A666Fa',
-        BridgeMapper: '0x9cb2820EA169D37aFa13C097776bDDB9b19d3C14'
-      }
+      fuse: defer(function () {
+        if (this.web3.network === 'mainnet') {
+          return {
+            HomeBridgeFactory: '0x93EF4d4032E053978aA71792Efd05d8b583a2B78',
+            BridgeMapper: '0x41063a48F46EE7E20E7EbAd0185992724B4Ee56c'
+          }
+        } else {
+          return {
+            HomeBridgeFactory: '0xFb5CC1688Ec06c57cbAB6cC34c33413154A666Fa',
+            BridgeMapper: '0x9cb2820EA169D37aFa13C097776bDDB9b19d3C14'
+          }
+        }
+      })
     }
   },
   mongo: {
